+++ ui/src/api/alarms.js
@@ -147,7 +147,8 @@ export async function fetchAlarmsWithInfos(simCardInstanceId, pagination) {
   `;
 
   const response = await query(queryStr);
-  if (!response.data) return;
+  
+  if (response.errors) return response;
 
   return response.data.alarmsWithInfo;
 }